pyproject.toml
[tool.poetry]
name = "inverno"
version = "0.1.0"
description = ""
authors = ["werew <werew@ret2libc.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
click = "^7.1.2"
yfinance = "^0.1.59"
pandas = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}
pylint = {version = "^2.7.4", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
inverno = "inverno.cli:main"

